SASS-10721 - Tech debt - Foreign Property : Rename route for foreign property endpoint
