+++ app/controllers/foreignproperty/ForeignPropertyJourneyAnswersController.scala
@@ -14,16 +14,17 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.foreignproperty
 
 import actions.AuthorisedAction
+import controllers.RequestHandler
 import errorhandling.ErrorHandler
 import models.common._
 import models.request.foreign._
 import models.request.foreign.adjustments.ForeignPropertyAdjustmentsWithCountryCode
-import models.request.foreign.sba.ForeignPropertySbaWithCountryCode
 import models.request.foreign.allowances.ForeignPropertyAllowancesWithCountryCode
 import models.request.foreign.expenses.ForeignPropertyExpensesWithCountryCode
+import models.request.foreign.sba.ForeignPropertySbaWithCountryCode
 import play.api.Logging
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import services.ForeignPropertyService
