+++ app/controllers/ukproperty/JourneyAnswersController.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.ukproperty
 
 import actions.AuthorisedAction
+import controllers.RequestHandler
 import errorhandling.ErrorHandler
 import models.RentalsAndRaRAbout
 import models.common._
@@ -30,7 +31,6 @@ import play.api.Logging
 import play.api.libs.json._
 import play.api.mvc.{Action, AnyContent, ControllerComponents, Result}
 import services.PropertyService
-import services.journeyAnswers.JourneyStatusService
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import utils.JsonSupport._
 
@@ -40,20 +40,12 @@ import scala.util.{Failure, Success}
 
 class JourneyAnswersController @Inject() (
   propertyService: PropertyService,
-  journeyStatusService: JourneyStatusService,
   auth: AuthorisedAction,
   cc: ControllerComponents
 )(implicit ec: ExecutionContext)
     extends BackendController(cc) with ErrorHandler with Logging with RequestHandler {
 
-  def fetchPropertyData(taxYear: TaxYear, nino: Nino, incomeSourceId: IncomeSourceId): Action[AnyContent] = auth.async {
-    implicit request =>
-      withJourneyContext(taxYear, incomeSourceId, nino, JourneyName.AllJourneys, request) { ctx =>
-        handleResponse(OK) {
-          propertyService.getFetchedPropertyDataMerged(ctx, nino, incomeSourceId)
-        }
-      }
-  }
+
 
   def savePropertyAbout(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino): Action[AnyContent] = auth.async {
     implicit request =>
@@ -306,41 +298,6 @@ class JourneyAnswersController @Inject() (
       }
     }
 
-  def setStatus(taxYear: TaxYear, incomeSourceId: IncomeSourceId, journeyName: String): Action[AnyContent] =
-    auth.async { implicit request =>
-      val ctx =
-        JourneyContext(taxYear, incomeSourceId, request.user.getMtditid, JourneyName.withNameInsensitive(journeyName))
-      val requestBody = parseBody[JourneyStatusData](request)
-
-      requestBody match {
-        case Success(validatedRes) =>
-          validatedRes.fold[Future[Result]](Future.successful(BadRequest)) {
-            case JsSuccess(value, _) =>
-              journeyStatusService.setStatus(ctx, value).value.map(_ => NoContent)
-            case JsError(err) =>
-              Future.successful(toBadRequest(CannotReadJsonError(err.toList)))
-          }
-        case Failure(err) => Future.successful(toBadRequest(CannotParseJsonError(err)))
-      }
-    }
-
-  def setForeignStatus(taxYear: TaxYear, incomeSourceId: IncomeSourceId, journeyName: String, countryCode: String): Action[AnyContent] =
-    auth.async { implicit request =>
-      val ctx =
-        JourneyContext(taxYear, incomeSourceId, request.user.getMtditid, JourneyName.withNameInsensitive(journeyName))
-      val requestBody = parseBody[JourneyStatusData](request)
-
-      requestBody match {
-        case Success(validatedRes) =>
-          validatedRes.fold[Future[Result]](Future.successful(BadRequest)) {
-            case JsSuccess(value, _) =>
-              journeyStatusService.setForeignStatus(ctx, value, countryCode).value.map(_ => NoContent)
-            case JsError(err) =>
-              Future.successful(toBadRequest(CannotReadJsonError(err.toList)))
-          }
-        case Failure(err) => Future.successful(toBadRequest(CannotParseJsonError(err)))
-      }
-    }
 
   def savRentalsAndRaRSBA(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino): Action[AnyContent] =
     auth.async { implicit request =>
