+++ build.sbt
@@ -25,9 +25,10 @@ lazy val coverageSettings: Seq[Setting[?]] = {
     "partials.*",
     "app.Routes",
     "ukAndForeignProperty.Routes",
+    "ukProperty.Routes",
     "health.Routes",
-    "foreign.Routes",
-    "foreign.RoutesPrefix",
+    "foreignProperty.Routes",
+    "foreignProperty.RoutesPrefix",
     "ukAndForeignProperty.RoutesPrefix",
 
   )
