--- conf/foreign.routes
@@ -1,17 +1,0 @@
-# Foreign Property specific routes #######################
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-select-country/:nino/answers           controllers.ForeignPropertyJourneyAnswersController.saveSelectCountrySection(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-tax/:nino/answers                      controllers.ForeignPropertyJourneyAnswersController.saveForeignPropertyTax(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-income/:nino/answers                   controllers.ForeignPropertyJourneyAnswersController.saveForeignIncome(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-expenses/:nino/answers                 controllers.ForeignPropertyJourneyAnswersController.saveForeignPropertyExpenses(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-allowances/:nino/answers               controllers.ForeignPropertyJourneyAnswersController.saveForeignPropertyAllowances(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-adjustments/:nino/answers              controllers.ForeignPropertyJourneyAnswersController.saveForeignPropertyAdjustments(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-sba/:nino/answers                      controllers.ForeignPropertyJourneyAnswersController.saveForeignPropertySba(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
-
-POST          /property/:taxYear/:incomeSourceId/foreign-property-delete/:nino/answers                   controllers.ForeignPropertyJourneyAnswersController.deleteForeignPropertyAnswers(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
