+++ conf/ukAndForeignProperty.routes
@@ -1,3 +1,3 @@
 ############### UK & Foreign Property ##############
 
-POST          /property/:taxYear/:incomeSourceId/uk-foreign-property-about/:nino/answers           controllers.UkAndForeignJourneyAnswersController.saveUkAndForeignPropertyAbout(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
+POST          /:taxYear/:incomeSourceId/uk-foreign-property-about/:nino/answers           controllers.ukandforeignproperty.UkAndForeignJourneyAnswersController.saveUkAndForeignPropertyAbout(taxYear: TaxYear, incomeSourceId: IncomeSourceId, nino: Nino)
